users & reservations テンプレート作成

+++ src/main/kotlin/com/kaiyujin/anemoneapi/reservations/Reservation.kt
@@ -1,3 +1,3 @@
 package com.kaiyujin.anemoneapi.reservations
 
-data class Reservation(val id: Int? = null, val firstName: String, val lastName: String)
+data class Reservation(val id: Long? = null, val firstName: String, val lastName: String)
